refactor(EMA): separate eds widgets to separate file
